requirements.txt
amqp==5.1.1
anyio==3.7.1
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
click==8.1.6
colorama==0.4.6
comm==0.1.4
cryptography==41.0.3
cssselect2==0.7.0
dataclasses-json==0.5.14
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
django-admin-rangefilter==0.10.0
django-admin-sortable2==2.1.9
django-awesome-standards==2020.6.5
django-ckeditor-5==0.2.9
django-ckeditor==6.7.0
django-codemirror2==0.2
django-constance==2.9.1
django-des==2.4.1
django-filer==3.0.4
django-filter==23.2
django-jinja==2.10.2
django-js-asset==2.1.0
django-model-utils==4.3.1
django-ok-seo==0.9.4
django-parler==2.3
django-polymorphic==3.1.0
django-postie==0.8.4
django-redis==5.3.0
django-rest-auth==0.9.5
django-robots==5.0
django-rosetta==0.9.9
django-script-pattern==0.3.3
django-solo==1.1.3
django-vuejs-translate==0.6.4
Django==4.2.4
djangorestframework-camel-case==1.1.2
djangorestframework==3.14.0
drf-orjson-renderer==1.1.3
easy-thumbnails==2.8.5
env==0.1.0
environ==1.0
exceptiongroup==1.1.2
executing==1.2.0
fake-useragent==1.2.1
fastjsonschema==2.18.0
fqdn==1.5.1
googlemaps==4.10.0
idna==3.4
ipykernel==6.25.1
ipython-genutils==0.2.0
ipython==8.14.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
jsmin==3.0.1
json5==0.9.14
jsonpointer==2.4
jsonschema-specifications==2023.7.1
jsonschema==4.19.0
jupyter-client==8.3.0
jupyter-console==6.6.3
jupyter-core==5.3.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter-server-terminals==0.4.4
jupyter-server==2.7.0
jupyter==1.0.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.24.0
jupyterlab-widgets==3.0.8
jupyterlab==4.0.4
kombu==5.3.1
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mistune==3.0.1
mock==5.1.0
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
notebook-shim==0.2.3
notebook==7.0.2
orjson==3.9.4
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
paypalrestsdk==1.13.1
pickleshare==0.7.5
Pillow==10.0.0
pip==23.1.2
platformdirs==3.10.0
polib==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
redis==4.6.0
referencing==0.30.2
reportlab==4.0.4
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
Send2Trash==1.8.2
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
svglib==1.5.1
terminado==0.17.1
tilda-wrapper-api==0.1.5
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing-extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
wheel==0.40.0
widgetsnbextension==4.0.8

